Cache page state in history by default
We do this for performance reasons and also so that AJAX navigation acts
more like regular static-page navigation.

When a new page is visited (appended to session history), the browser
will save the scroll position on the previous page. If the back button is
clicked, that scroll position is restored.  Unfortunately, javascript
cannot modify the scroll position before the popstate event is fired, so
the following would occur:

- back clicked
- scroll position restored
- popstate
- page.read() called (async)
- wait for data
- page.render() (sync)

which was unsightly because the scroll jumps before the content changes.
With caching, the sequence goes:

- back clicked
- scroll position restored
- popstate
- page.render() (sync)

and there is no time for the jump. This is more like efficient
static-page browsing, where the browser caches the HTML and then simply
recalls it on back.

Do not be afraid of caching too much stuff in the History API, because
the standard says the user agent should cull older entries first if any
sort of buffer limit is reached (i.e. history is a FIFO buffer).